Fixed a problem discovered when trying to implement a js solution for 303520:
Problem with Java <-> Javascript date conversions that were based on milliseconds. And the bug is not noticeable for common years - in my case made a simple code that would do what the user wants not work correctly because user inputs 7 July 0010 and in javascript when you would say date.getUTCDate() you would get 5 instead of 7.
This code should work after the fix for the problem in this case:
		var year = newValue.getFullYear();
		if (year < 60) {
			year += 2000;
		} else if (year < 100) {
			year += 1900;
		}
		if (year != newValue.getFullYear())	{
			newValue.setFullYear(year);
			date_plain = newValue;
		}
Before the fix it would transofm 7 July 0010 into 5 July 2010...